runtests_requirements.txt
-r requirements.txt

psycopg2
mysqlclient
django-redis
python-memcached
pylibmc
pytz
Jinja2

django-debug-toolbar
beautifulsoup4
